refactor MovieDetailPage component
