+++ project-movie-app/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './index.css'
 import { App } from './App.js'
-import { MovieDetailPage } from './MovieDetailPage/MovieDetailPage.js';
+import { MovieDetailPage } from './MovieDetailPage/MovieDetailPage';
 
 const root = document.getElementById("root")!;
